Secure product modification and profile views

+++ products/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
+
 from .models import Product, Category
 from .forms import ProductForm
 
@@ -68,9 +70,14 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store"""
     products = Product.objects.all()
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do this')
+        return redirect(reverse('home'))
     
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -92,8 +99,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """Edit a product in DH Bakes store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do this')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -116,8 +128,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete a product in the DH Bakes store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do this')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'{product.name} successfully deleted')
